+++ python/cugraph/__init__.py
@@ -26,6 +26,8 @@ from cugraph.community import (
 from cugraph.structure import (
     Graph, 
     DiGraph, 
+    MultiGraph,
+    MultiDiGraph,
     renumber, 
     symmetrize, 
     symmetrize_df