+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ## Bug Fixes
 - PR #634 renumber vertex ids passed in analytics
 - PR #649 Change variable names in wjaccard and woverlap to avoid exception
+- PR #651 fix cudf error in katz wrapper and test nstart
 
 # cuGraph 0.11.0 (11 Dec 2019)