+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - PR #629 get rid of gdf_column in pagerank 
 - PR #641 Add codeowners
 - PR #646 Skipping all tests in test_bfs_bsp.py since SG BFS is not formally supported
-
+- PR #652 Remove gdf_column in BFS
 ## Bug Fixes
 - PR #634 renumber vertex ids passed in analytics
 - PR #649 Change variable names in wjaccard and woverlap to avoid exception